# backend/wwwroot/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "rainbow-unicorn",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Customer/getCustomerDetails/{customerName}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "Get customer details",
        "parameters": [
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Customer/getCustomerSolutions/{customerName}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "Get all the solutions that a Customer subscribed to.",
        "parameters": [
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Customer/getStocks/{customerName}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "Get all the stocks that a Customer added to their dashboard.",
        "parameters": [
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/CustomerStock/{ticker}/{customerName}/{interval}/{stockName}": {
      "post": {
        "tags": [
          "CustomerStock"
        ],
        "summary": "Adds a stock to customer dashboard.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/CustomerStock": {
      "put": {
        "tags": [
          "CustomerStock"
        ],
        "summary": "Update interval for a customer's stock.",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerStock"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerStock"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerStock"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerStock"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/CustomerStock/{ticker}/{customerName}": {
      "delete": {
        "tags": [
          "CustomerStock"
        ],
        "summary": "Remove stock from the customer's dashboard.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerStock"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Solution": {
      "get": {
        "tags": [
          "Solution"
        ],
        "summary": "Get all existing solutions by GS",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Solution/{solutionId}": {
      "get": {
        "tags": [
          "Solution"
        ],
        "summary": "Retrieve solution details base on given ID",
        "parameters": [
          {
            "name": "solutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Solution/AllId": {
      "get": {
        "tags": [
          "Solution"
        ],
        "summary": "Returns an array of all solution IDs",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solution"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/api/Stock/{ticker}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stock"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stock"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stock"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreditCard": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "expiryDate": {
            "type": "string",
            "nullable": true
          },
          "cvv": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        },
        "additionalProperties": false
      },
      "Customer": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "passwordSalt": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "solutionCustomers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SolutionCustomer"
            },
            "nullable": true
          },
          "creditCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditCard"
            },
            "nullable": true
          },
          "userPurchases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPurchase"
            },
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "nullable": true
          },
          "customerStocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerStock"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomerStock": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "interval": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "stock": {
            "$ref": "#/components/schemas/Stock"
          }
        },
        "additionalProperties": false
      },
      "Feature": {
        "type": "object",
        "properties": {
          "featureTitle": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "solutionId": {
            "type": "integer",
            "format": "int32"
          },
          "solution": {
            "$ref": "#/components/schemas/Solution"
          }
        },
        "additionalProperties": false
      },
      "Solution": {
        "type": "object",
        "properties": {
          "solutionId": {
            "type": "integer",
            "format": "int32"
          },
          "solutionTitle": {
            "type": "string",
            "nullable": true
          },
          "organisation": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "new": {
            "type": "boolean"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "nullable": true
          },
          "solutionCustomers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SolutionCustomer"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SolutionCustomer": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "solutionId": {
            "type": "integer",
            "format": "int32"
          },
          "datePurchased": {
            "type": "string",
            "format": "date-time"
          },
          "amountPayable": {
            "type": "number",
            "format": "float"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "solution": {
            "$ref": "#/components/schemas/Solution"
          }
        },
        "additionalProperties": false
      },
      "Stock": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "nullable": true
          },
          "stockName": {
            "type": "string",
            "nullable": true
          },
          "customerStocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerStock"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "userPurchases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPurchase"
            },
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        },
        "additionalProperties": false
      },
      "UserPayment": {
        "type": "object",
        "properties": {
          "purchaseId": {
            "type": "string",
            "nullable": true
          },
          "paymentNumber": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "fulfilled": {
            "type": "boolean"
          },
          "userPurchase": {
            "$ref": "#/components/schemas/UserPurchase"
          }
        },
        "additionalProperties": false
      },
      "UserPurchase": {
        "type": "object",
        "properties": {
          "purchaseId": {
            "type": "string",
            "nullable": true
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "numberOfPayments": {
            "type": "integer",
            "format": "int32"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userPayments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPayment"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "apiKey",
        "description": "Standard Authorization header using the Bearer Scheme (\"bearer {token}\")",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}